refactor: separated Certificate handling to its own gateway and mapper pair
