+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/infrastructure/jpa/makedecision/MakeDecisionGatewayMapper.java
@@ -2,13 +2,10 @@ package uk.gov.justice.laa.dstew.access.infrastructure.jpa.makedecision;
 
 import java.time.Instant;
 import java.util.Map;
-import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import uk.gov.justice.laa.dstew.access.domain.ApplicationDomain;
-import uk.gov.justice.laa.dstew.access.domain.CertificateDomain;
 import uk.gov.justice.laa.dstew.access.entity.ApplicationEntity;
-import uk.gov.justice.laa.dstew.access.entity.CertificateEntity;
 import uk.gov.justice.laa.dstew.access.entity.DecisionEntity;
 import uk.gov.justice.laa.dstew.access.entity.MeritsDecisionEntity;
 import uk.gov.justice.laa.dstew.access.entity.ProceedingEntity;
@@ -17,38 +14,6 @@ import uk.gov.justice.laa.dstew.access.model.MeritsDecisionStatus;
 
 /** Converts domain records ↔ JPA entities for operations in the makeDecision use case. */
 public class MakeDecisionGatewayMapper {
-
-  /**
-   * Maps a {@link CertificateEntity} to a {@link CertificateDomain}.
-   *
-   * @param certificateEntity the JPA certificate entity
-   * @return the domain
-   */
-  public CertificateDomain toCertificateDomain(CertificateEntity certificateEntity) {
-    return CertificateDomain.builder()
-        .id(certificateEntity.getId())
-        .applicationId(certificateEntity.getApplicationId())
-        .certificateContent(certificateEntity.getCertificateContent())
-        .createdBy(certificateEntity.getCreatedBy())
-        .updatedBy(certificateEntity.getUpdatedBy())
-        .build();
-  }
-
-  /**
-   * Maps a {@link CertificateDomain} to a new {@link CertificateEntity}. Use only for INSERT (id is
-   * null). For UPDATE, use the load-and-apply pattern.
-   *
-   * @param certificateDomain the domain
-   * @return a new entity for persistence
-   */
-  public CertificateEntity toCertificateEntity(CertificateDomain certificateDomain) {
-    CertificateEntity entity = new CertificateEntity();
-    entity.setApplicationId(certificateDomain.applicationId());
-    entity.setCertificateContent(certificateDomain.certificateContent());
-    entity.setCreatedBy(certificateDomain.createdBy());
-    entity.setUpdatedBy(certificateDomain.updatedBy());
-    return entity;
-  }
 
   /**
    * Mutates the managed {@link ApplicationEntity} in-place from the supplied domain. Does NOT
@@ -64,11 +29,13 @@ public class MakeDecisionGatewayMapper {
    */
   public void applyDecisionToEntity(ApplicationEntity entity, ApplicationDomain domain) {
     DecisionEntity decisionEntity =
-        Optional.ofNullable(entity.getDecision()).orElseGet(() -> DecisionEntity.builder().build());
+        entity.getDecision() != null ? entity.getDecision() : DecisionEntity.builder().build();
+
     decisionEntity.setOverallDecision(
         domain.decision().overallDecision() != null
             ? DecisionStatus.valueOf(domain.decision().overallDecision().name())
             : null);
+
     decisionEntity.setModifiedAt(domain.decision().modifiedAt());
     entity.setDecision(decisionEntity);
 
@@ -76,9 +43,11 @@ public class MakeDecisionGatewayMapper {
     entity.setModifiedAt(Instant.now());
 
     if (entity.getProceedings() != null && domain.proceedings() != null) {
+
       Map<UUID, ProceedingEntity> proceedingEntityMap =
           entity.getProceedings().stream()
               .collect(Collectors.toMap(ProceedingEntity::getId, p -> p));
+
       domain
           .proceedings()
           .forEach(
@@ -86,8 +55,9 @@ public class MakeDecisionGatewayMapper {
                 ProceedingEntity proceedingEntity = proceedingEntityMap.get(proceedingDomain.id());
                 if (proceedingEntity != null && proceedingDomain.meritsDecision() != null) {
                   MeritsDecisionEntity meritsDecision =
-                      Optional.ofNullable(proceedingEntity.getMeritsDecision())
-                          .orElseGet(MeritsDecisionEntity::new);
+                      proceedingEntity.getMeritsDecision() != null
+                          ? proceedingEntity.getMeritsDecision()
+                          : new MeritsDecisionEntity();
                   meritsDecision.setDecision(
                       proceedingDomain.meritsDecision().decision() != null
                           ? MeritsDecisionStatus.valueOf(
